feat(prisma): move PrismaClient to config/prisma.ts
