+++ src/config/prisma.ts
@@ -1,0 +1,5 @@
+import { PrismaClient } from "../generated/prisma/index.js";
+
+const prisma = new PrismaClient();
+
+export default prisma;
