+++ src/middlewares/graphqlAuth.ts
@@ -2,7 +2,7 @@ import jwt from "jsonwebtoken";
 
 import HttpError from "../utils/httpError.js";
 
-import { prisma } from "../index.js";
+import prisma from "../config/prisma.js";
 import { redisClient } from "../config/redis.js";
 
 import AuthenticatedRequest from "../types/authenticatedRequest.js";
